+++ src/popup.js
@@ -1,29 +1,11 @@
 
 export function displaypopup() {
     let support = "<a href=\"https://ko-fi.com/sdoehren\" target=\"_blank\" rel=\"noopener noreferrer\">\n" +
-        "                    <img src=\"https://img.shields.io/badge/ko--fi-Support-red?style=flat-square&amp;logo=ko-fi\" alt=\"ko-fi\"></a>\n" +
-        "                <a href=\"https://www.patreon.com/sdoehren?utm_source=Tension+Pool&utm_medium=Referral&utm_campaign=Module+Popup\" target=\"_blank\" rel=\"noopener noreferrer\">\n" +
-        "                    <img src=\"https://img.shields.io/badge/Patreon-Support-red?style=flat-square&amp;logo=patreon\" alt=\"Patreon\"></a>\n"
-
+        "                    <img src=\"https://img.shields.io/badge/ko--fi-Support-red?style=flat-square&amp;logo=ko-fi\" alt=\"ko-fi\"></a>"
 
     let post = "<h1>Tension Pool</h1>\n" + support +
-        "<h1 id=\"change-log\">Change Log</h1>\n" +
-        "<h4 id=\"0-0-50\">0.0.50</h4>\n" +
-        "<ul>\n" +
-        "<li>Protection against overfilling the pool has been added.</li>\n" +
-        "</ul>\n" +
-        "<h4 id=\"0-0-47-macro-breaking-change\">0.0.47 - Macro Breaking Change</h4>\n" +
-        "<ul>\n" +
-        "<li>Tension die short code had been changed from dt6 to just dt to fix an issue with compatibility with Dice So Nice</li>\n" +
-        "</ul>\n" +
-        "<h4 id=\"0-0-46\">0.0.46</h4>\n" +
-        "<ul>\n" +
-        "<li>d100 option added</li>\n" +
-        "<li>Option to Trigger a macro when a complication occurs added</li>\n" +
-        "<li>Minor Visual Improvements</li>\n" +
-        "<li>Clicking the pool will now add a die</li>\n" +
-        "<li>Dice So Nice 4 Support</li>\n" +
-        "</ul>\n"
+        "<h1 id=\"Large Change\">Large Change</h1>\n" +
+        "<p>Please beaware that Scene Controls have been removed and replaced with </p>"
 
 
     let d = new Dialog({
@@ -34,11 +16,6 @@ export function displaypopup() {
                 icon: '<i class="fas fa-check"></i>',
                 label: "Close",
                 callback: () => game.settings.set("tension-pool", "DontShowAgain", true)
-            },
-            three: {
-                icon: '<i class="fab fa-patreon"></i>',
-                label: "Goto Patreon",
-                callback: () => window.open("https://www.patreon.com/sdoehren", '_blank')
             }
         },
         default: "one",